# api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/epoch/current": {
			"get": {
				"operationId": "GetCurrentEpoch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Epoch"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/health": {
			"get": {
				"operationId": "GetHealthStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										}
									},
									"required": [
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"V1 Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/mempool/size": {
			"get": {
				"operationId": "GetMempoolSize",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Mempool"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/pool/distribution": {
			"get": {
				"operationId": "GetPoolDistribution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Pool"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/pool/timing": {
			"get": {
				"operationId": "GetPoolTiming",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Pool"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/tx": {
			"get": {
				"operationId": "GetTxList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/tx/{hash}": {
			"get": {
				"operationId": "GetTxDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "hash",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tx/confirmation": {
			"get": {
				"operationId": "GetConfirmationDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "hash",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			}
		},
		"/api/v1/tx/confirmed": {
			"get": {
				"operationId": "GetConfirmedTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pool",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/tx/stats": {
			"get": {
				"operationId": "GetTxStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tx/timing": {
			"get": {
				"operationId": "GetTxTiming",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"V1 Transaction"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v2/health": {
			"get": {
				"operationId": "GetHealthStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										}
									},
									"required": [
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"V2 Health"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}